move updater security previews screen to be after Wi-Fi setup
When opting into security previews, the Updater will do a new check on server. If the user has a SIM
inserted, it might start downloading over that. Move the security previews screen to after
Wi-Fi setup. This way, we can prefer to download over Wi-Fi if the user has set it up.

Of course, if the user setting up on an outdated build (where the newer build number is not a
security preview release), the updater might download the newer build over SIM data regardless.